+++ src/triangle.h
@@ -15,7 +15,7 @@ typedef struct {
 typedef struct {
     vec2_t points[3];
     uint32_t color;
-    float avg_depth;
+    double avg_depth;
 } triangle_t;
 
 #endif
